date_histogram handler functions refactored and moved to /lib folder
