+++ public/response_handlers/lib/utils.js
@@ -1,0 +1,5 @@
+function containsAgg(obj, aggId) {
+  return Object.keys(obj).includes(aggId);
+}
+
+export { containsAgg };
